0226 - responsiveness fixes-refactoring-dependencies fix and script for win/mac
