+++ package.json
@@ -9,6 +9,7 @@
         "@dnd-kit/utilities": "^3.2.1",
         "@reduxjs/toolkit": "^1.7.1",
         "@tanstack/react-query": "^4.32.0",
+        "@testing-library/jest-dom": "^6.1.4",
         "axios": "1.6.0",
         "caniuse-lite": "^1.0.30001517",
         "classnames": "^2.3.1",
@@ -41,7 +42,6 @@
         "@types/react-redux": "^7.1.24",
         "@types/react-router-dom": "^5.3.3",
         "@types/react-router-hash-link": "^2.4.5",
-        "@types/testing-library__jest-dom": "^6.0.0",
         "@types/testing-library__react": "^10.2.0",
         "@types/uuid": "^8.3.1",
         "@typescript-eslint/eslint-plugin": "^6.4.0",
@@ -58,7 +58,7 @@
     },
     "scripts": {
         "start": "set PORT=3100 && react-scripts start",
-        "start_mac": "PORT=3100 react-scripts start",
+        "startmac": "PORT=3100 react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject",
@@ -66,7 +66,7 @@
         "coins": "json-server --watch data/coins.json --port 5008 --delay 1000",
         "utilities": "json-server --watch data/utilities.json --port 5006 --delay 1000",
         "startall": "concurrently \"npm run utilities\" \"npm run bubbles\" \"npm run coins\" \"npm start\"",
-        "startall_mac": "concurrently \"npm run utilities\" \"npm run bubbles\" \"npm run coins\" \"npm start_mac\"",
+        "startallmac": "concurrently \"npm run utilities\" \"npm run bubbles\" \"npm run coins\" \"npm run startmac\"",
         "lint": "eslint src/*.{ts,tsx}",
         "lint:fix": "eslint --fix src/*.{ts,tsx}",
         "format": "prettier --write src/**/*.{ts,tsx,md,scss} --config ./.prettierrc",
