+++ src/features/building-story/ManageLevels/ManageLevels.module.scss
@@ -1,0 +1,10 @@
+@use "css/z-index" as z;
+
+.controls {
+    width: 400px;
+    @include z.z-index();
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 5px;
+}
