+++ src/features/memoryGame-story/Game/GameEnd/GameEnd.module.scss
@@ -3,7 +3,6 @@ $textColor: crimson;
 .finished {
     width: 100%;
     height: 100%;
-    min-width: 180px;
     background-color: rgba(9, 77, 9, 0.637);
     display: flex;
     align-items: center;
